+++ pyvela/pyvela/spnta.py
@@ -678,9 +678,10 @@ class SPNTA:
             10. `param_units.txt` - Parameter unit strings (astropy format)
             11. `param_scale_factors.txt` - Scale factors that convert parameter values from Vela's internal units to 'normal' units
             12. `param_true_values.txt` - Parameter values used for simulation, taken from the 'truth' par file
-            13. `prior_info.json` - Prior distributions on all free parameters (JSON format)
-            14. `prior_evals.npy` - Prior distributions evaluated in the posterior range for plotting (numpy format)
-            15. `summary.json` - Information about the machine, environment, sampler, and input (JSON format)
+            13. `param_autocorr.txt` - Parameter autocorrelation lengths (from the thinned chains)
+            14. `prior_info.json` - Prior distributions on all free parameters (JSON format)
+            15. `prior_evals.npy` - Prior distributions evaluated in the posterior range for plotting (numpy format)
+            16. `summary.json` - Information about the machine, environment, sampler, and input (JSON format)
         """
         samples = self.rescale_samples(samples_raw)
 
@@ -700,12 +701,16 @@ class SPNTA:
         )
 
         self.save_resids(params_median, f"{outdir}/residuals.txt")
+        param_autocorr = emcee.autocorr.integrated_time(
+            samples_raw, quiet=True, has_walkers=False
+        )
 
         np.savetxt(f"{outdir}/param_default_values.txt", self.default_params)
         np.savetxt(f"{outdir}/param_names.txt", self.param_names, fmt="%s")
         np.savetxt(f"{outdir}/param_prefixes.txt", self.param_prefixes, fmt="%s")
         np.savetxt(f"{outdir}/param_units.txt", self.param_units, fmt="%s")
         np.savetxt(f"{outdir}/param_scale_factors.txt", self.scale_factors)
+        np.savetxt(f"{outdir}/param_autocorr.txt", param_autocorr)
 
         if truth_par_file is not None:
             np.savetxt(