+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Unreleased
 ## Added
+- Can restart emcee from a prior chain
+- Script/function to resample existing chain
+- Saving autocorrelation lengths of chains
 ## Changed
+- Normalize the prior plot in `pyvela-plot` to improve visibility
 ## Fixed
 - Mark Julia lines wrongly missed in coverage with COV_EXCL_LINE to make the coverage % more accurate
 ## Removed
@@ -25,7 +29,6 @@
 - Save the full prior information along with the results
 - Convert `RNAMP` and `RNIDX` to `TNREDAMP` and `TNREDGAM`
 - Plot priors in `pyvela-plot`
-- Can restart emcee from a prior chain
 ## Changed
 - Residual plot in `pyvela-plot` script
 - Made GP noise marginalization the default in `SPNTA`
@@ -33,6 +36,7 @@
 - Residual plot in `pyvela-plot` script
 - Made GP noise marginalization the default in `SPNTA`
 - Move `info_dict()`, `save_new_parfile()`, `save_resids()` from `pyvela_script` to `SPNTA`
+- Prior plotting is now normalized
 ## Fixed
 - Julia example script `run_example.jl`
 - In `_gls_lnlike_serial`, return  -inf if `Σinv` is not positive definite.